Added account delete

+++ frontend/cpuDiffApp/src/app/menu/menu.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { AuthService } from '../auth.service';
 
 @Component({
   selector: 'app-menu',
@@ -7,12 +8,18 @@ import { Component } from '@angular/core';
 })
 export class MenuComponent {
 
+  constructor(private authService: AuthService){}
+
   handleClickLogOut(): void {
     localStorage.clear();
   }
 
   handleClickDeleteAccount(): void {
-    localStorage.clear();
+    const confirmation = window.confirm("Czy na pewno chcesz usunąć konto?");
+    if (confirmation) {
+      this.authService.deleteAccount().subscribe(() => {localStorage.clear();}, (error) => {console.error(error)});
+      localStorage.clear();
+    }
   }
 
 }
